Added cleanup script and refactor common code
