+++ e2e_samples/fabric_dataops_sample/devops/azure-pipelines-ci-qa.yml
@@ -52,9 +52,9 @@ stages:
 
           - script: |
               echo "Running tests in $(WORKING_DIR)/libraries/test..."
-              pytest libraries
+              pytest .
             displayName: 'Run Tests'
-            workingDirectory: $(WORKING_DIR)/
+            workingDirectory: $(WORKING_DIR)/libraries
 
   - stage: BuildFabric
     jobs:
